solving soft delete issue Your entity uses @DeleteDateColumn for deletedAt. TypeORM will not return rows where deletedAt is set (not null) when using find() or findOneBy() by default.

+++ src/products/products.service.ts
@@ -204,6 +204,10 @@ export class ProductsService {
 
    public async deleteproduct(id: number) {
     const product = await this.getsingleProducts(id);
+    if (!product)
+      throw new NotFoundException(`product not found ${id}`, {
+        description: 'the product does not deleted',
+      });
     await this.productsRepository.remove(product);
     return { message: 'product was deleted' };
    }
